requirements.txt
Django==3.1.13
djangorestframework==3.13.1
drf-spectacular==0.21.1
psycopg2==2.9.3
redis==4.1.2
celery==5.2.3
django-celery-beat==2.2.1
djangorestframework-simplejwt==5.0.0
django_seed==0.3.1